creates separate template files
